Перепиши реализацию меню и фильтров
